[상진규] refactor : interface 분기
